Disable controller spans by default
