+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/configuration/Configuration.java
@@ -196,7 +196,7 @@ public class Configuration {
     // ignoreRemoteParentNotSampled is currently needed
     // because .NET SDK always propagates trace flags "00" (not sampled)
     public boolean ignoreRemoteParentNotSampled = true;
-    public boolean captureControllerSpans = true;
+    public boolean captureControllerSpans;
     // this is just here to detect if using this old setting in order to give a helpful message
     @Deprecated public boolean httpMethodInOperationName;
     public LiveMetrics liveMetrics = new LiveMetrics();
