+++ test/smoke/appServers/global-resources/controller_spans_enabled_applicationinsights.json
@@ -5,6 +5,6 @@
     "instance": "testroleinstance"
   },
   "preview": {
-    "captureControllerSpans": false
+    "captureControllerSpans": true
   }
 }
