+++ test/smoke/testApps/SpringCloudStream/src/smokeTest/java/com/microsoft/applicationinsights/smoketest/SpringCloudStreamControllerSpansEnabledTest.java
@@ -32,80 +32,84 @@ import com.microsoft.applicationinsights.smoketest.schemav2.RequestData;
 import java.util.List;
 import org.junit.Test;
 
-@UseAgent("controller_spans_disabled")
-public class JmsControllerSpansDisabledTest extends AiSmokeTest {
+@UseAgent("controller_spans_enabled")
+@WithDependencyContainers({
+  @DependencyContainer(
+      value = "confluentinc/cp-zookeeper",
+      portMapping = "2181",
+      environmentVariables = {"ZOOKEEPER_CLIENT_PORT=2181"},
+      hostnameEnvironmentVariable = "ZOOKEEPER"),
+  @DependencyContainer(
+      value = "confluentinc/cp-kafka",
+      portMapping = "9092",
+      environmentVariables = {
+        "KAFKA_ZOOKEEPER_CONNECT=${ZOOKEEPER}:2181",
+        "KAFKA_ADVERTISED_LISTENERS=PLAINTEXT://${CONTAINERNAME}:9092",
+        "KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR=1"
+      },
+      hostnameEnvironmentVariable = "KAFKA")
+})
+public class SpringCloudStreamControllerSpansEnabledTest extends AiSmokeTest {
 
   @Test
   @TargetUri("/sendMessage")
   public void doMostBasicTest() throws Exception {
     List<Envelope> rdList = mockedIngestion.waitForItems("RequestData", 2);
 
-    Envelope rdEnvelope1 = getRequestEnvelope(rdList, "GET /sendMessage");
+    Envelope rdEnvelope1 = rdList.get(0);
     String operationId = rdEnvelope1.getTags().get("ai.operation.id");
     List<Envelope> rddList =
         mockedIngestion.waitForItemsInOperation("RemoteDependencyData", 2, operationId);
     assertEquals(0, mockedIngestion.getCountForType("EventData"));
 
-    Envelope rdEnvelope2 = getRequestEnvelope(rdList, "message process");
-    Envelope rddEnvelope1 = getDependencyEnvelope(rddList, "message send");
-    Envelope rddEnvelope2 = getDependencyEnvelope(rddList, "GET /");
+    Envelope rdEnvelope2 = rdList.get(1);
+    Envelope rddEnvelope1 = rddList.get(0);
+    Envelope rddEnvelope2 = rddList.get(1);
 
     RequestData rd1 = (RequestData) ((Data<?>) rdEnvelope1.getData()).getBaseData();
     RequestData rd2 = (RequestData) ((Data<?>) rdEnvelope2.getData()).getBaseData();
-
     RemoteDependencyData rdd1 =
         (RemoteDependencyData) ((Data<?>) rddEnvelope1.getData()).getBaseData();
     RemoteDependencyData rdd2 =
         (RemoteDependencyData) ((Data<?>) rddEnvelope2.getData()).getBaseData();
 
+    if (!rdd1.getName().equals("GreetingsController.sendMessage")) {
+      RemoteDependencyData rddTemp = rdd1;
+      rdd1 = rdd2;
+      rdd2 = rddTemp;
+
+      Envelope rddEnvelopeTemp = rddEnvelope1;
+      rddEnvelope1 = rddEnvelope2;
+      rddEnvelope2 = rddEnvelopeTemp;
+    }
+
     assertEquals("GET /sendMessage", rd1.getName());
     assertTrue(rd1.getProperties().isEmpty());
     assertTrue(rd1.getSuccess());
 
-    assertEquals("message send", rdd1.getName());
+    assertEquals("GreetingsController.sendMessage", rdd1.getName());
     assertNull(rdd1.getData());
-    assertEquals("Queue Message | jms", rdd1.getType());
-    assertEquals("message", rdd1.getTarget());
+    assertEquals("InProc", rdd1.getType());
+    assertNull(rdd1.getTarget());
     assertTrue(rdd1.getProperties().isEmpty());
     assertTrue(rdd1.getSuccess());
 
-    assertEquals("message process", rd2.getName());
-    assertEquals("message", rd2.getSource());
-    assertTrue(rd2.getProperties().isEmpty());
-    assertTrue(rd2.getSuccess());
-
-    assertEquals("GET /", rdd2.getName());
-    assertEquals("https://www.bing.com", rdd2.getData());
-    assertEquals("Http", rdd2.getType());
-    assertEquals("www.bing.com", rdd2.getTarget());
+    assertEquals("greetings send", rdd2.getName());
+    assertNull(rdd2.getData());
+    assertEquals("Queue Message | kafka", rdd2.getType());
+    assertEquals("greetings", rdd2.getTarget());
     assertTrue(rdd2.getProperties().isEmpty());
     assertTrue(rdd2.getSuccess());
 
+    assertEquals("greetings process", rd2.getName());
+    assertEquals("greetings", rd2.getSource());
+    assertTrue(rd2.getProperties().isEmpty());
+    assertTrue(rd2.getSuccess());
+    assertTrue(rd2.getMeasurements().containsKey("timeSinceEnqueued"));
+
     assertParentChild(rd1, rdEnvelope1, rddEnvelope1, "GET /sendMessage");
+    assertParentChild(rdd1, rddEnvelope1, rddEnvelope2, "GET /sendMessage");
     assertParentChild(
-        rdd1.getId(), rddEnvelope1, rdEnvelope2, "GET /sendMessage", "message process", false);
-    assertParentChild(
-        rd2.getId(), rdEnvelope2, rddEnvelope2, "message process", "message process", false);
-  }
-
-  private static Envelope getRequestEnvelope(List<Envelope> envelopes, String name) {
-    for (Envelope envelope : envelopes) {
-      RequestData rd = (RequestData) ((Data<?>) envelope.getData()).getBaseData();
-      if (rd.getName().equals(name)) {
-        return envelope;
-      }
-    }
-    throw new IllegalStateException("Could not find request with name: " + name);
-  }
-
-  private static Envelope getDependencyEnvelope(List<Envelope> envelopes, String name) {
-    for (Envelope envelope : envelopes) {
-      RemoteDependencyData rdd =
-          (RemoteDependencyData) ((Data<?>) envelope.getData()).getBaseData();
-      if (rdd.getName().equals(name)) {
-        return envelope;
-      }
-    }
-    throw new IllegalStateException("Could not find dependency with name: " + name);
+        rdd2.getId(), rddEnvelope2, rdEnvelope2, "GET /sendMessage", "greetings process", false);
   }
 }
